requirements.txt
streamlit
pandas
chromadb
sentence-transformers
google-generativeai
crewai
more-itertools
